pdftoraster filter converted to filter function
