+++ cupsfilters/filter.h
@@ -223,6 +223,13 @@ extern int pdftops(int inputfd,
 		   void *parameters);
 
 
+extern int pdftoraster(int inputfd,
+				int outputfd,
+				int inputseekable,
+				filter_data_t *data,
+				void* parameters);
+
+
 extern int pstops(int inputfd,
 		  int outputfd,
 		  int inputseekable,
